// src/app/Models/Intern/intern-detail.ts
export class InternDetail {
  'internId': number;
  'name': string;
  'emailID': string;
  'mobileNo': string;
  'qualification': string;
  'department': string;
  'internshipAvailability': string;
  'internshipDuration': string;
}

export class InternQuestionAnswer {
  'internId': number;
  'question': string;
  'answer': string;
}